% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnglm-package.R
\docType{package}
\name{learnglm-package}
\alias{learnglm-package}
\alias{learnglm}
\title{Tutorials to learn the General Lake Model}
\description{
Tutorials to learn the General Lake Model
}
\author{
\email{stachel2@msu.edu}
}
